refactor: Path -> FilePath, enum으로 관리
