+++ weekly-mission/src/main/java/org/prgrms/weeklymission/customer/repository/BlackCustomerRepository.java
@@ -7,12 +7,12 @@ import org.springframework.stereotype.Repository;
 import java.io.*;
 import java.util.*;
 
-import static org.prgrms.weeklymission.utils.Path.BLACK_CUSTOMER_FILE_DB_PATH;
+import static org.prgrms.weeklymission.utils.FilePath.BLACK_CUSTOMER_FILE_DB_PATH;
 
 @Repository
 @Slf4j
 public class BlackCustomerRepository implements CustomerRepository {
-    private static final String path = BLACK_CUSTOMER_FILE_DB_PATH;
+    private static final String path = BLACK_CUSTOMER_FILE_DB_PATH.getPath();
     private final File file = new File(path);
     private final BufferedWriter output = new BufferedWriter(new FileWriter(file));
     private final BufferedReader input = new BufferedReader(new FileReader(file));
