--- weekly-mission/src/main/java/org/prgrms/weeklymission/utils/Path.java
@@ -1,8 +1,0 @@
-package org.prgrms.weeklymission.utils;
-
-public class Path {
-    private Path() {}
-
-    public static final String VOUCHER_FILE_DB_PATH = "/Users/kinseonghyeon/Desktop/springboot-basic/weekly-mission/db/voucher-db.txt";
-    public static final String BLACK_CUSTOMER_FILE_DB_PATH = "/Users/kinseonghyeon/Desktop/springboot-basic/weekly-mission/db/black-db.csv";
-}
